Add UI components for AI Application Generator and enhance prompt reformulation
