+++ src/config/constants.py
@@ -1,6 +1,6 @@
 import os
 
 API_MODEL = "gpt-4o-mini"
-API_TEMPERATURE = 0.2
+API_TEMPERATURE = 0.4
 MAX_TOKENS_DEFAULT = 4000
 MAX_TOKENS_LARGE = 8000
