+++ src/config/prompts.py
@@ -285,4 +285,8 @@ REFORMULATION_PROMPT = """
     Do NOT add major new features that weren't implied in the original.
     
     Return ONLY the reformulated description, without any explanations or metadata.
+    Additionally, this project requires that ALL necessary files are generated by AI as part of the architecture. 
+    Include instructions to generate every required configuration file (like requirements.txt, package.json), 
+    documentation files (README.md), and any other supporting files. No templates should be used - everything 
+    should be derived from the requirements.
     """
